refactor: organize starlark convert functions to new go file
